+++ docs/include/samplecode/SoundTouchClient/CreateZoneFromDevices.py
@@ -4,7 +4,7 @@ from bosesoundtouchapi.models import *
 try:
     
     # create SoundTouch device instance.
-    device:SoundTouchDevice = SoundTouchDevice("192.168.1.131") # Bose SoundTouch 10
+    device:SoundTouchDevice = SoundTouchDevice("192.168.1.81") # Bose SoundTouch 10
             
     # create SoundTouch client instance from device.
     client:SoundTouchClient = SoundTouchClient(device)
@@ -14,8 +14,8 @@ try:
     print("\nZone Status Before:\n%s" % zoneBefore.ToString(True))
 
     # create new device instances for all zone members.
-    device_master:SoundTouchDevice = SoundTouchDevice("192.168.1.131") # master
-    device_member:SoundTouchDevice = SoundTouchDevice("192.168.1.130") # member
+    device_master:SoundTouchDevice = SoundTouchDevice("192.168.1.81") # master
+    device_member:SoundTouchDevice = SoundTouchDevice("192.168.1.80") # member
             
     # create a new master zone configuration on the device.
     masterZone:Zone = client.CreateZoneFromDevices(device_master, [device_member])