+++ docs/include/samplecode/SoundTouchClient/SetUserTrackControl.py
@@ -10,7 +10,7 @@ try:
     # and advance to the next track.
 
     # create SoundTouch device instance.
-    device:SoundTouchDevice = SoundTouchDevice("192.168.1.131") # Bose SoundTouch 10
+    device:SoundTouchDevice = SoundTouchDevice("192.168.1.81") # Bose SoundTouch 10
             
     # create SoundTouch client instance from device.
     client:SoundTouchClient = SoundTouchClient(device)